Move TaxYear definition
