+++ src/components/CreatePDF.tsx
@@ -3,9 +3,8 @@ import { Helmet } from 'react-helmet'
 import { usePager } from './pager'
 import Alert from '@material-ui/lab/Alert'
 import { useSelector } from 'react-redux'
-import { Information, Asset, State } from 'ustaxes/core/data'
+import { Information, Asset, State, TaxYear } from 'ustaxes/core/data'
 import { YearsTaxesState } from 'ustaxes/redux'
-import { TaxYear } from 'ustaxes/data'
 
 import yearFormBuilder from 'ustaxes/forms/YearForms'
 import { isLeft, runAsync } from 'ustaxes/core/util'
