+++ src/components/SummaryData.ts
@@ -1,5 +1,5 @@
 import Form from 'ustaxes/core/irsForms/Form'
-import { TaxYear } from 'ustaxes/data'
+import { TaxYear } from 'ustaxes/core/data'
 import F1040For2020 from 'ustaxes/forms/Y2020/irsForms/F1040'
 import F1040For2021 from 'ustaxes/forms/Y2021/irsForms/F1040'
