+++ src/components/YearStatusBar.tsx
@@ -2,7 +2,7 @@ import { Link } from '@material-ui/core'
 import { ReactElement, useState } from 'react'
 import { useSelector } from 'react-redux'
 import { YearsTaxesState } from 'ustaxes/redux'
-import { TaxYears } from 'ustaxes/data'
+import { TaxYears } from 'ustaxes/core/data'
 import YearDropDown from './YearDropDown'
 
 const YearStatusBar = (): ReactElement => {
