+++ src/components/payments/EstimatedTaxes.tsx
@@ -1,9 +1,8 @@
 import { ReactElement } from 'react'
 import { FormProvider, useForm } from 'react-hook-form'
 import { usePager } from 'ustaxes/components/pager'
-import { EstimatedTaxPayments } from 'ustaxes/core/data'
+import { EstimatedTaxPayments, TaxYear } from 'ustaxes/core/data'
 import { YearsTaxesState } from 'ustaxes/redux'
-import { TaxYear } from 'ustaxes/data'
 import { Currency, LabeledInput } from 'ustaxes/components/input'
 import { Patterns } from 'ustaxes/components/Patterns'
 import { FormListContainer } from 'ustaxes/components/FormContainer'
