+++ src/core/data/index.ts
@@ -1,5 +1,13 @@
 import { enumKeys } from '../util'
 
+export enum TaxYears {
+  Y2019 = 2019,
+  Y2020 = 2020,
+  Y2021 = 2021
+}
+
+export type TaxYear = keyof typeof TaxYears
+
 export enum PersonRole {
   PRIMARY = 'PRIMARY',
   SPOUSE = 'SPOUSE',
