+++ src/redux/migration.ts
@@ -1,5 +1,5 @@
 import { enumKeys } from 'ustaxes/core/util'
-import { TaxYears } from 'ustaxes/data'
+import { TaxYears } from 'ustaxes/core/data'
 import { blankState } from './reducer'
 import { USTState } from './store'
