+++ src/tests/arbitraries.ts
@@ -2,9 +2,8 @@ import * as fc from 'fast-check'
 import * as util from 'ustaxes/core/util'
 import * as arbitraries from 'ustaxes/core/tests/arbitraries'
 import { YearsTaxesState } from 'ustaxes/redux'
-import { TaxYear, TaxYears } from 'ustaxes/data'
 import prand from 'pure-rand'
-import { Asset, AssetType } from 'ustaxes/core/data'
+import { Asset, AssetType, TaxYear, TaxYears } from 'ustaxes/core/data'
 
 export const taxYear: fc.Arbitrary<TaxYear> = fc.constantFrom(
   ...util.enumKeys(TaxYears)
