+++ src/tests/common/YearsStatusBarMethods.tsx
@@ -7,7 +7,7 @@ import {
 } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import { enumKeys } from 'ustaxes/core/util'
-import { TaxYear, TaxYears } from 'ustaxes/data'
+import { TaxYear, TaxYears } from 'ustaxes/core/data'
 import DomMethods from './DomMethods'
 
 export default class YearStatusBarMethods extends DomMethods {
